Refactor authentication components by creating a new LoginForm component to streamline login processes for Auth0, Supabase, and Basic Auth; update LoginPage and AuthGuard to utilize the new component for improved code organization and maintainability.
